Pygame-RPG 21: Creating a QuestManager
This PR is focused on creating a manager for quests. To begin to do
this, we define a quest as a dictionary and store the information in
JSON to be retrieved at the start of the game.

A quest has the following structure:
`Quest Type`: A string that indicates what type of quest it is

`Next Quest`: A string that has the dictionary key of the next quest in
the chain. If there is not next quest in the quest chain, it is empty
string.

`Target`: A dictionary that has the objective of the quest in dictionary
form. For example, if the objective is the get 5 potions, the dictionary
would be {"Potion": 5}.

`Context`: A string that is only used in "Arrive" type quests, it
displays what the context needs to be to complete the quest. If it isn't
an "Arrive" type quest, then it is empty string.

`Range`: A list of 2 x coordinates that indicate the values that the
player has to be between in order to complete the quest. For example, if
the range is [500, 600], the player needs to have an x coordinate
between the two number so 501 to 599 is fair game. This is only used in
"Arrive" type quests.

`Reward`: This is a dictionary, formatted the same way as lootpool in
BattleManager, that has the rewards that the player gets from completing
the quest.

To effectively manage quests, a manager class was created.

`QuestManager`: This class manages quests and keeps track of factors
that affect quest progression. It also holds a list of active quests and
completed quests, this is how progression can be viewed. This class has
the following attributes, `questDict`, `knight`, `activeQuest`,
`completedQuests`, `enemiesKilled`, `npcsInteractedWith`.

`questDict`: A reference to the master list of quests.

`knight`: reference to the knight object

`activeQuests`: a list of quests that the player is trying to complete

`completedQuests`: a list of quests that the player has completed

`enemiesKilled`: a dictionary that pairs the name of an enemy to the
number of times the player has killed them (string: int)

`npcsInteractedWith`: A dictionary that pairs the name of an NPC and the
number of times the player has interacted with them.

The class has the following functions, `add_quest()`,
`get_all_keys_from_quest_targets()`, `remove_unnecessary_keys()`,
and `check_for_completion()`.

`add_quest()`: This function takes in a string and adds the associated
quest to the activeQuest list if the string key exists in the master
dictionary.

`get_all_keys_from_quest_targets()`: This function loops through all
active quests and gets the keys from their `Target` dictionary and
returns them in the form of a list.

`remove_unnecessary_keys()`: This function deletes any irrelevant keys
from both `enemiesKilled` and `npcsInteractedWith`. An irrelevant key is
a key that isn't mentioned in any active quest.

`check_for_completion()`: This function takes in a string and an
integer. These two values are used for "Arrive" type quests to see if
the player meets the conditions for completing that quest. This function
loops through all active quests and determines if they are completed. If
the function determines that the quest is completed, it checks if
there is a next quest (if there is it is added to the activeQuest list,
gives out the rewards of the quest and adds it to the list of completed
quests. If the quest hasn't been completed, it stays in the active quest
list.

Other Changes:
- SaveManager now saves QuestManager details as well
- A script was created to add new quests to the master dictionary
- save files were updated to include QuestManager details
- files using SaveManager were updated to include QuestManager in the
  init function
- Added a "Quest" key to the events JSON for future integration