Pygame-RPG 21.1 Updating EventManager
This change is to update event manager to process "Quest" type events.
This includes adding a reference to the QuestManager to the `init()`
function for the EventManager class. The `process_events()` function was
also changed to include code that processes "Quest" type events.

The JSON and the other test files that include the EventManager object
were also updated due to this change. The other notable thing is that
the `get_event_dict()` function is no longer a class function but a
normal function.