Pygame-RPG 21.1 Updating EventManager
This change is to update event manager to process "Quest" type events.
This includes adding a reference to the QuestManager class in the
EventManager class. Because of this, the `process_events()` function was
changed to include code that processes "Quest" type events now using
the QuestManager object.

The JSON for these test events were added to the project and the other
class test files that include the EventManager object
were also updated due to this change. The`get_event_dict()` function is no
longer a class function but a normal function because it doesn't serve
any class specific function.