Pygame-RPG 21.2 Preparing for animation work
This PR focuses on preparing the entity objects for the work that will
be done using the sprites. We do this by both adding in a new folder for
the enemy sprites as well as creating the `Sprite` attribute for Entity
objects which points to the folder which contains their sprites.

Because of these changes, both the Knight class and the Enemy class need
to update their `init()` functions to include the new `sprite`
attribute. Due to this, the save files were also updated.