Pygame-RPG 21.3 Simplifying code using Sprite class
This PR is focused on simplifying the code by updating the `Knight`
class so it can work as a sprite class.

To do this, the `Knight` class has to inherit from the `Sprite` class.
However, since `Enemy`, another Entity class, will also need to inherit
from the sprite class. To make life easier, `Entity` is the class that
will inherit from `Sprite` which will then get inherited by `Knight` and
`Enemy` objects.

Because we will now be using the `Sprite` class, there are three key
changes that were made to `Knight`. The first change was that the class
now needs to track it's own coordinates via an `x` and `y` attribute.

The second change was that the class would now need to track its own
animations. This is done using three new attributes `aniStatus` which is
used to represent what animation is in use, `anitracker` a number that
increases with each frame and determines which frame of the animation we
are on and `maxAniVal` which is the highest possible frame number. This
keeps the animation from trying to access a frame that doesn't exist.
For example, if a walking cycle has 11 frames, `maxAniVal` would be 11
and the code wouldn't try to access a 12 frame because of this. These
three are all inherited from `Entity`.

The final change was the addition of the actual attributes used in
animating, the surface and rect objects from Pygame. `image` is a
surface that we load by using the `Sprite`, `Name`, `aniStatus` and `aniTracker`
attributes to get. `Sprite` points to the folder which has the
animations for the object. The `Name` and `aniStatus` attributes refer
to the animation that is in use while `aniTracker` determines the frame
in use. Once we have our image, we use the `get_rect()` function to set
the `rect` attribute and center the rect using the `x` and `y`
attributes. There is also now a `flipped` attribute in the `Knight`
class that is a boolean indicator for flipping the image so that it
faces the opposite direction.

The above are largely logic that will eventually move from the `Knight`
class up to the `Entity` class, just so I don't have to recode it for
the `Enemy` class.

The next section is about the functions that were added or majorly changed.
For the `Entity` class, it received two new functions
`get_max_animation_val()` and `reset_max_animation_val()`.

`get_max_animation_val()`: This function returns a number representing
the largest possible frame based off of the `Sprite`, `Name` and `aniStatus`
attributes.

`reset_max_animation_val()`: this function simply sets the `maxAniVal`
attribute to the returned value of `get_max_animation_val()`.

For the `Knight` class, it received 3 new functions,
`set_image_and_rect()`, `update()` and `equals`.

`set_image_and_rect()`: This function simply updates the `image` and
`rect` attributes whenever the time to change the animation comes (this
is handled externally from the class). if the `flipped` attribute is
true, the `flip()` method in pygame is called and the output  is set to
the `image` attribute resulting in a flipped image.

`update()`: This is a function that determines how our animation works.
In this case, it increments the `aniTracker` by 1 every time we call it
and once it reaches a multiple of 10, some criteria is checked for how
the animation is to proceed. If the image needs to be updated, a flag
called `update` is set to True and the `set_image_and_rect()` function
is called to update the `image` and `rect` attributes.

`equals()`: This function takes another `Knight` class object and
compares them based off of certain dictionary keys. If the values of the
specific keys match, the objects are considered the same and True is
returned, otherwise, false is returned. This function is
meant to be a replacement for overriding the `__eq__` function. Because
of how the `Sprite` class works, overriding this specific method
resulted in an error.

The `load_dict()` function also received some updates as a result of
these changes. This is due to both `image` and `rect` not being JSON
serializable, as such, we can't just call `vars()` and store them and
just set the dictionaries to each other. Instead, we store SOME of
knight's attributes, things like the stats, animation state, etc. and
derive the rest from these values. After updating the dictionary, we
call `reset_max_animation_val()` to reset the max animation and
`set_image_and_rect()` to generate the image and rectangle needed for
animation. We also reset the types of some things like ensuring that
`Status` is a tuple and `moveList` is a list.

The next set of changes are aimed at leveraging the `Sprite` class to
simplify the animations. This change is in the `Rpg2.py`, we remove
almost all of the animation work that was done externally for the
`Knight` object and we simplify the remaining functions. functions like
`handle_basic_input()` were trimmed down since the logic needed for it
was unnecessary with the `Sprite` class

A sprite group was also created to facilitate animating more than one
sprite object but for now, the group only consists of the `Knight`
object.

all references to the `x`, `animationTracker`, `spot` variables were
removed. Any function that referenced `x` and the `Knight` object were
updated to use the the `x` attribute of the object instead.

Other things:
- As always, many tests broke and needed to be updated to handle this new
  logic. Namely, everything that has to do with `Save_Manager` broke.
- Changed the name of the `Enemy_Sprites` folder to be `Entity_Sprites`
  and added the Knight animations there.
- Updated the scripts to reflect the change in the folder name
- Updated the range of some things to accommodate for the animation
  change
- Changed how the `change_screen()` function in both `Rpg2.py` and the
  `ScreenManager` class work moving forward
- Save files updated
- Created a sorting function for the `get_max_animation_val()` function
  called `sort_func`