Pygame-RPG 21.4 Revamping Animations with Sprites
This commit consists of changes regarded the sprites. Instead of
manually adjusting the size of the sprites using some external software,
an alternative way of doing the same thing was available using pygame.
Using `pygame.transform.scale()`, we are able to scale the size of the
sprites to a good size for the game WITHOUT losing quality (which was a
big issue).

This commit also adds more sprites for the Knight class (most of which
are not in use yet). Because using `pygame.transform.scale()` is so
effective, it was used for the portraits of the Knight class (one for
dialogue and the other for displaying the health and mana bars). As
such, the other pngs that were for that task were deleted.

Other Changes:
- Some slight changes to ScreenManager JSON dictionary regarding the
  range for triggering the treasure chest animation
- Added more documentation