Pygame-RPG 21.5 Housekeeping and bugfixing
This change mostly consists of dealing with warnings from the Pycharm IDE.
Things like spacing between functions etc. The important changes are the
following.

`DialogueManager`: The `load_portrait()` function received a change to
fix a bug in the code. The code failed to remove the name portion of a
string before returning it so that was added. The function also returns
a string that is stripped of whitespace as well. Because of this, the
tests regarding this function needed to be tweaked.

`BattleManager`: the `get_move_dict()` and `get_status_dict()` functions
are no longer class functions but static functions. The reason for this
is that it's convenient to be able to pull these functions over when
necessary rather than having to re-implement them every time.

`Enemy`: The `super()` function is now called for inheritance instead of
creating a new Enemy object. This is just cleaner and matches Knight's
implementation. This is in preparation for `Enemy` objects to be
animated.

Also, since we now use scaling, two portrait PNGs for Knight were
deleted. As a result, the `UI_Manager_draw.py` file needed to be updated
to use a scaled up version of the Knight's portrait. as a result, the
math for where it is meant to be drawn was also changed.