Pygame-RPG 21.6 Fixing NPC Animations
This PR is focused on implementing Sprite class functionality for NPCs.
In order to do this, 3 changes needed to be made. The first change was a
proper structure for NPCs that could be converted to JSON. Second, an
NPC class and an NPC management system (NPCManager class). Finally, the
above needs to be implemented into the `Rpg2.py` file.

The schema for an NPC is as follows:
- `Name`: The name of the NPC object, this is used to get their
  animations.
- `Sprite`: A path to the folder containing the NPCs animations
- `Scale`: A tuple containing the desired size for the NPC sprite
- `Dialogue`: This is a list of keys for dialogue/quest events
- `Context`: This is a string that indicates which contexts the NPC is
  supposed to appear in
- `Pos`: The coordinates in which the NPC is supposed to be drawn on

All of this is in the "JSON/NPCs/NPCs.json" file.

Next, an `NPC` class was created. This class inherits from the `Sprite`
class and has no really unique values. `aniStatus`, `aniTracker`,
`maxAniVal`, `image`, `rect` are the attributes that aren't mentioned in
the schema but these aren't unique attributes as they are shared with
the other `Sprite` classes like `Knight`.

In fact, many functions are recycled from the `Knight` class. The `NPC`
class has simplified versions of the `set_image_and_rect()` function
(which doesn't allow sprites to be flipped in this version), `update()`
(since the NPCs only have an idle animation it doesn't need to consider
the edge cases in the `Knight` version). It also has a direct copy of
the `get_max_animation_val()` function from `Knight` along with it's
sorting function. This is inefficient but it will do for now.

The only unique functions that the `NPC` class has are the following,
`add_event_keys()` and `get_event_key()`.

`add_event_keys()`: This function takes in a list of string event keys
and directly adds it to the `Dialogue` list. This addition of the two
lists maintains order so the integrity of the queue remains. This
function is meant to update the dialogue of the NPCs after major story
events.

`get_event_key()`: This function pops and returns the first item in the
`Dialogue` list if the list isn't empty. If the list IS empty, a generic
dialogue key is returned instead.

Next, the class meant to manage the NPCs, `NPCManager`. This is the one
manager class not part of the managers package because it's kind of a
pain if it was. The class has more to do with Entities than on screen
management so I figure it works. This class has the following
attributes, `NPCDict`, `NPCGroup`, `knight`.

`NPCDict`: This is a reference to the dictionary of the NPCs.

`NPCGRoup`: This is a `Group` object for NPCs

`knight`: reference to the `Knight` object

The functions for this class are the following, `create_NPC()`,
`get_NPCs()`, `update_NPC()`, `get_colliding()`,
`get_interaction_event()`.

`create_NPC()`: This function takes in a key for the NPC dictionary and
uses the dictionary information to create and return an NPC object

`get_NPCs()`: This function takes in a string representing the current
screen and first removes all the current NPCs in `NPCGroup` before
finding and creating and adding the NPCs whose
`context` attribute matches the given string to `NPCGroup`.

`update_NPC()`: This function takes in a name and a list of events.
If there is an NPC in the group whose `Name` attribute matches the given
name, the events are added to it via the `add_event_keys()` function.

`get_colliding()`: This function determines if the `Knight` object is
colliding with an `NPC` in the group. If so, the `NPC` is returned, if
not then None is returned

`get_interaction_event()`: This function takes in an pygame event list.
If there is a sprite that is "colliding" with the `Knight` object and
the player is pressing the up arrow key to interact, the latest event
key for the NPC is returned using the `get_event_key()` function.
Otherwise, None is returned.

Other Changes:
- added and renamed files in the event_text folder
- tweaked how `load_portrait` and `draw_dialogue()` work in `DialogueManager`
- updated the scripts
- added new events
- made some simplifications to `Knight` class and added more comments
- Created NPCs.json file
- Added the animations and portrait for the Gossiper NPC
- deleted the previous code for managing NPCs and implemented the
  current solution into Rpg2.py