Pygame-RPG 21.8 Housekeeping and bugfixing
This PR is about changing the names of the class variables so that they
don't mirror the name of their class. This is really only a problem for
the following classes. `NPCManager`, `UIManager` and `UIHandler`.

Because of this, the names of the object for those classes will be changed
to the following.

`NPCManager` -> `npcManager`
`UIManager` -> `uiManager`
`UIHandler` -> `uiHandler`

This way, it will be obvious when we are referring to the class and when
we're referring to the object.