Pygame-RPG 21.9 Housekeeping and bugfixing
This PR focuses on optimizing the functions in the `Getters.py` file.
Since most of the functions in the file repeat the same steps to get
retrieve a dictionary from a given JSON file, a new function was created
to simplify this and reduce redundancy. This function is called
`get_json_from_file_path()`.

`get_json_from_file_path()`: This function takes in a file path and
loads a dictionary with the stored JSON information. The dictionary is
returned and ensures that the file is properly closed.

Now, the getter functions for the JSON dictionaries simply returns the
output of `get_json_from_file_path()` with a specific file path. For
example, `get_quest_dict()` would pass in the `JSON/Quests/Quests.json`
path to the function.

Next was optimizing the `Scripts.py` file. Using the newly improved
`Utils` package, the scripts can be further streamlined to make changes
to the structure of the project easier. The `fill-saves` script was also
updated to actually work now since it was missing some objects
beforehand. Some variable names were also changed to make it more
obvious what their role was.