Pygame-RPG 21.10 Creating Object Animations
This PR is focused on creating a system for managing objects in the
game, whether they be doors, chests etc. As always, we need to create a
schema for what defines an object. We also need to define a schema for
certain object types.

The schema for an Object is as follows:

`Name`: The key for the object

`Object Type`: This is a key for the `Object_def.json` file that defines
the animation information.

`Flipped`: Determines whether the image should be flipped or not.

`Context`: The background that this Object appears in

`Pos`: The x and y coordinates for the object

`Events`: A list of event keys associated with the Object

The schema for the type of an object is as follows:

`Sprite`: The location of the object's sprites

`Scale`: The scale the image should be set to

Next, we need to define the object class.

The object class has all of the attributes in both of the object
dictionaries, this includes `Name`, `ObjectType`, `Sprite`, `Scale`,
`Flipped`, `Context`, `Pos`, `Events`. All of these values are taken
from a dictionary passed to the constructer. This dictionary does have
two notable entries not mentioned in the object JSON files. That is,
`aniStatus`, and `aniTracker`. The reason for this is that unlike NPCs
or the player/enemy objects, the objects NEED to remember their
animation and stick to it. (we wouldn't want a chest to re-open
every time you load back in).

This class also has the `get_max_animation_val()` function as well as
the `set_image_and_rect_function()`, `update()` and `get_event_key()`
function.

`get_max_animation_val()`: This function is identical to the other
sprite classes except the `ObjectType` attribute is passed as the second
argument instead of `Name`. This is because the object type is the
indicator for the animations, not `Name`.

`set_image_and_rect()`: Like I mentioned in the above function, the file
path replaces the `self.Name` portion with `self.ObjectType` as that is
the proper indicator. It also supports flipping images like with
`Knight`'s version of the function.

`update()`: Almost a 1-1 copy of `Knight`'s `update()` function except
it fixes a small bug in how the Entity classes update. This bug will be
addressed later. Also, the condition for keeping the animation stuck at
the last animation is different. Instead of checking if the `aniStatus`
attribute is "Dead", it checks if the `aniStatus` is "Opening"

`get_event_key()`: This function is identical to the `get_event_key()`
function in the `NPC` class. The only difference is that the generic key
returned depends on the `ObjectType` attribute

After defining the `Object` class, we also need a manager for these
objects to make everything seamless.

The `ObjectManager` class has the following attributes, `objectDict`,
`objectGroup`, and `knight`.

`objectDict`: This is a reference to a dictionary with the "complete"
Object information. This is gotten by calling the
`get_complete_object_dict()`. this function essentially merges the
information from the two object dictionaries found in `Object_def.json`
and `Objects.json` and adds the two keys `aniStatus` and `aniTracker` to
create all the information needed to make an `Object`.

`objectGroup`: a Sprite group for the Objects

`knight`: reference to the `Knight` object

The class also has the following functions, `create_object()`,
`save_and_empty()`, `get_objects()`, `get_colliding()`,
`get_interaction_event()`.

`create_object()`: Takes in a key for an object and returns the Object
object.

`save_and_empty()`: saves the information for each sprite in the group
by updating the `Events`, `aniStatus` and `aniTracker` information in
the given object's dictionary. It uses the `Name` attribute to find the
correct dictionary. Also, if the `anitracker` attribute is too large,
we update the object's animation to re-adjust it.

`get_objects()`: works the same as the one in the `NPCManager` class,
just adjusted for this class. Takes in a context and returns the objects
that match the given context.

`get_colliding()`: This function determines which sprite is "colliding"
with the player character and returns that Object. However, unlike the
same function in the `NPCManager` class, the collision is testing if the
knight object's rectangle is colliding with the mid right point of the
Object.

`get_interaction_event()`: If there is an Object colliding with the
player and the player presses the up arrow key the event key of the
object is returned. If the object has the type "Treasure_Chest" it's
animation status is changed to "Opening" and the `aniTracker` and
`maxAniVal` attributes are updated as well. If there is no Object
colliding with the player character, None is returned.

To accomodate these changes, some scripts to the `script` and
`Scripts.py` files. The `add-object` and `add-object-def` options.
`Utils` also got 3 new functions `get_object_def_dict()`,
`get_object_dict()` and `get_complete_object_dict()`:

`get_object_def_dict()`: This is a getter function for retrieving the
dictionary stored in `Object_def.json`.

`get_object_dict()`: This is a getter function for retrieving the
dictionary stored in `Objects.json`

`get_complete_object_dict()`: This retrieves the information in the
`get_object_def_dict()` and `get_object_dict()`, it then combines the
information in both of those dictionaries using the `Object Type` to add
in the `Sprite` and `Scale` key from the Object definition dictionary
along with creating the new keys `aniStatus` and `aniTracker`. It then
returns the new dictionary with all of these keys.

Other Changes:
- Added the JSON files `Object_def.json` and `Objects.json`
- added the sprites for the "Treasure_Chest" object type
- removed the `animationTracker3` and `spot3` references from `Rpg2.py`
  file and replaced it by integrating the `ObjectManager` class in it's
  place.
- added a new event to `Events.json`